requirements.txt
python-dotenv # For reading environment variables stored in .env file
langchain
langchain-community
unstructured # Document loading
langchain-huggingface 
sentence_transformers==3.4.0
chromadb # Vector storage
langchain-chroma>=0.1.2
# install markdown depenendies with: `pip install "unstructured[md]"` after install the requirements file. Leave this line commented out. 
